Refactor CLI to have separate commands
This patch will introduce a refactor/split of our CLI module. The
intention here is to make sure that each (sub)command will be treated
independently and have their own logic.

Currently, it is being divided into two categories:
* history
* query

The history subcommand will handle everything that is related to the
user history of conversation.

The query subcommand will handle anything that is related to user
queries. This is also the default subcommand.