Refactor the build specfile (#56)
In this patch, we refactored how we did the build for the CLA project.
This was needed, because of a couple of reasons:

- We were not including the submodules under `command_line_assistant` source package

- The binaries we were shipping were not working properly with root
-- Any time that we tried to execute the binary through `$ c`, it tried
to load from `lib64` instead of `lib`

To achieve those modifications, we created a new folder called `bin`,
where it has `c` and `clad`.

We also modified the pyproject.toml to include the submodules using the
Find Directive.