Add local history management and improve the command
Now we can save/view and query (basic) the history locally. The history
module allows for modularity to switch between LocalHistory and a future
implementation of a RemoteHistory that will be stored in the servers.

We also improved the `history` command to show in the terminal the
conversation history by having the following modes:
- All history
- First history
- Last history

More modes will be implemented in the future, but for now, this is enough.