Reverse order of the terminal output reader (#170)
This patch aims in reversing the order of the last output read from the
terminal to make the interface more friendly to the user. Instead of
having to think about positive and negative number ordering (python
indexes), they can just use positive number ordering and that will be
translated internally to the list.