Add ability to load credentials from systemd
This patch introduces a mechanism for loading the database credentials
from systemd credentials. This allow the user to specify their
database username/password via systemd-creds, and the daemon will load
and use it.

It will only work if the username/password is empty in the config file.
If it is detected, we will use what is in the config.