Limit down the size of message that will be sent to the backend (#189)
We don't want to allow big messages to go to the backend as of now. RAG
is very difficult to get it right, so we decided to trim the message
down to 2 KB (2048 bytes). Any message over it (combining all sources)
will be trimmed down to not overload the rag searches.