Add PlainText, PlainMarkdown and StaticSpinner renderers
PlainTextRenderer and PlainMarkdownRenderer subclass TextRenderer and MarkdownRenderer respectively. They disregard any decorators that might be added to them by other code and print plain text or raw, unformatted markdown as the case may be. StaticSpinnerRenderer subclasses SpinnerRenderer but writes a single "frame" to the stream instead of a series of animated frames.

The various renderer factory functions (create_text_renderer and create_markdown_renderer) have been updated with a keyword argument of 'plain' with a default value of false (preserving existing behavior). If plain is true, the factory function returns a Plain subclass instead of the rich-formatted class. The rest of the render flow remains unmodified. The various calls to create_text_renderer and create_markdown_renderer function have been updated to pass in the value of 'args.plain' wherever possible. This will enable plain rendering when the --plain flag is passed at the commandline. The --plain flag is also moved up from the chat subcommand to the base argument parser. This enables 'plain' functionality in history and feedback too.