Check for the presence of .bashrc.d
When installing the bash snippet that enables the interactive keybinding, check for the presence of ".bashrc.d" in the user's bash rc files. If no mention of the directory is found, a warning is printed, informing the user that shell integration requires the use of a snippet that loads files from ~/.bashrc.d.