Fix legal message regression found in 0.4
We found a regression for the legal message that it was being printed every time and was not respecting the getppid rule for the legal state file.

That regression got introduced due to the recent merge of #297 to main, where we added a --plain argument to prevent most of customizations to the output text. Since the legal message was being handled as part of a decorator in the rendering module, we broke the way the legal message was being controlled as we made the --plain default option.

This patch removes that decorator (no need to make such operation as part of rendering), and introduce it as a simple function that has less potential to break.